some changes realted to CLI
